feat: add SIP-013 SFT support (#74)
* chore: api db contract logs cursor

* feat: scan the api db for sft mint events

* feat: process token job sft

* feat: sft endpoints

* feat: detect sft mints through monitor

* docs: expand some comments [skip ci]

* fix: change token_count and token_number columns to pg numeric

+++ migrations/1670264425574_smart-contracts.ts
@@ -31,7 +31,7 @@ export function up(pgm: MigrationBuilder): void {
       notNull: true,
     },
     token_count: {
-      type: 'int',
+      type: 'numeric',
     },
     created_at: {
       type: 'timestamp',
